Introduce color extraction feature for web pages
This commit adds a new feature to extract colors from web pages. It includes the implementation of JavaScript functions for extracting, sorting and displaying colors in a chrome extension popup. New files for the feature include JavaScript, HTML and CSS for the popup, as well as a README. The manifest of the extension was also created.