Refactor src/js/colors.js and extend functionality of src/popup.js
In src/js/colors.js, unnecessary code comments are removed and function descriptions and input annotations are added to clarify the purpose of each function. In src/popup.js, buttons for copying values and HTML have been added, plus a mechanism to copy the image to the clipboard. Various helper functions are added to improve functionality. HTML is modified accordingly to incorporate these changes.