Add color copy function and improve UI responsiveness
Added a 'copyColor' function that allows users to copy color values directly from the UI. Also made improvements to the UI responsiveness by modifying the '.hidden' class and adding a pointer cursor to the '.colored-div'. The 'divContent' variable is now used to set the innerHTML content, which improves code readability and maintainability. The 'divContent' is also hidden during the execution of the 'EyeDropper' function to enhance user experience.