Update version, add design mode toggle and improve UI
This commit updates the 'Extract Colors DevTool' and 'extract-colors-extension' from version 1.0.2 to 1.0.3. It incorporates a new function to toggle the design mode of the document in 'colors.js' and implements the related button in 'popup.js', 'popup.html', and 'popup.css'. The UI is also improved; the color report generated now gets displayed as colored chips, providing a sleek and better visualization.