Modify advertisement cleaning criteria in cleaner.js
The list of targeted advertising classes in the cleaner.js file has been revised. The 'iframe' has been moved to another line and 'video' is now targeted as well along with a new class 'StickyFooter'. This enhances the ad cleaning function, helping to remove more potential ads from the content.