Migrate ruler CSS to JavaScript file and update version
The css related to the ruler functionality in the css/inject/ruler.css was migrated into js/inject/ruler.js as a string. The 'innerHTML' property was added to manage HTML embedded in JS. Additionally, the extension's version was updated in package.json and src/manifest.json. Please note that, css/inject/ruler.css was removed, and as a result, the related content_scripts entry was deleted from src/manifest.json. Finally, renamed 'system-info.js' to 'ruler-listener.js' to better describe its functionality.