Refactor screen capture and improve UI stylesheet
The screen capture functionality has been improved by refining the manner in which screen arrangements are calculated and handled. Optimized the HTML and CSS to better conform to UI designs, removing unnecessary margins and adjusting padding, in addition to slight changes to font sizes. A check has also been added to skip an operation if an array of filenames is empty.